+++ hitablelist.h
@@ -12,7 +12,7 @@ class hitable_list: public hitable {
 	{
 		list = l;
 	}
-	virtual bool hit(const ray& r, float t_min, float t_max, hit_record& rec) const;
+	virtual bool hit(const ray& r, double t_min, double t_max, hit_record& rec) const;
 	virtual bool bounding_box(aabb& box) const;
 
 	std::vector<hitable*> list;
