+++ mtl.h
@@ -7,13 +7,13 @@
 
 class Mtl {
 	public:
-		float Ns;
+		double Ns;
 		vec3 Ka;
 		vec3 Kd;
 		vec3 Ks;
 		vec3 Ke;
-		float Ni;
-		float d;
+		double Ni;
+		double d;
 		size_t illum;
 };
