+++ onb.h
@@ -17,7 +17,7 @@ class onb
 		vec3 v() const;
 		vec3 w() const;
 
-		vec3 local(float a, float b, float c) const;
+		vec3 local(double a, double b, double c) const;
 		vec3 local(const vec3& a) const;
 		void build_from_w(const vec3&);
