+++ ray.h
@@ -17,15 +17,15 @@ public:
 
 	vec3 origin() const { return A; }
 	vec3 direction() const { return B; }
-	vec3 point_at_parameter(float t) const {
+	vec3 point_at_parameter(double t) const {
 		return A + t * B;
 	}
 
 	vec3 A;
 	vec3 B;
-	float central_wl;
-	float min_wl;
-	float max_wl;
+	double central_wl;
+	double min_wl;
+	double max_wl;
 };
